user validation with HandlerInterceptor

+++ stockMate/src/main/java/com/stockm8/persistence/WarehouseDAO.java
@@ -10,7 +10,7 @@ public interface WarehouseDAO {
 	public void insertWarehouse(WarehouseVO wVO) throws Exception;
 	
 	// 회사정보를 통한 창고ID 조회
-	List<WarehouseVO> selectWarehousesByBusinessId(Long businessId) throws Exception;
+	List<WarehouseVO> selectWarehousesByBusinessId(Integer businessId) throws Exception;
 	
 	// 창고 목록 유효성 조회
     boolean existsById(int warehouseId, int businessId) throws Exception;
