+++ stockMate/src/main/java/com/stockm8/service/WarehouseService.java
@@ -10,5 +10,5 @@ public interface WarehouseService {
 	public void createWarehouse(WarehouseVO wVO) throws Exception;
 	
 	// 회사 ID를 통한 창고ID 정보 
-	List<WarehouseVO> getWarehousesByBusinessId(Long businessId) throws Exception;
+	List<WarehouseVO> getWarehousesByBusinessId(Integer businessId) throws Exception;
 }
